Display energy network capacity as FloatingLong

+++ src/api/java/mekanism/api/transmitters/IGridTransmitter.java
@@ -39,7 +39,7 @@ public interface IGridTransmitter<ACCEPTOR, NETWORK extends DynamicNetwork<ACCEP
 
     ITextComponent getTransmitterNetworkBuffer();
 
-    double getTransmitterNetworkCapacity();
+    int getTransmitterNetworkCapacity();
 
     @Nonnull
     FloatingLong getCapacityAsFloatingLong();
