+++ src/main/java/mekanism/common/transmitters/Transmitter.java
@@ -75,9 +75,8 @@ public abstract class Transmitter<ACCEPTOR, NETWORK extends DynamicNetwork<ACCEP
     }
 
     @Override
-    public double getTransmitterNetworkCapacity() {
+    public int getTransmitterNetworkCapacity() {
         //This isn't *fully* accurate as the fluid and gas networks only actually support up to max int currently
-        //TODO: FIXME - the energy network isn't as Double
         return hasTransmitterNetwork() ? getTransmitterNetwork().getCapacity() : getCapacity();
     }
